refactor: move getPendingDeposits to poaBridgeService; add return types
