+++ src/services/defuseBalanceService.ts
@@ -1,9 +1,7 @@
 import type { providers } from "near-api-js"
 import type { BaseTokenInfo } from "../types/base"
 import type { IntentsUserId } from "../types/intentsUserId"
-import { tokenAccountIdToDefuseAssetId } from "../utils/tokenUtils"
 import { batchBalanceOf } from "./intentsContractService"
-import { getDepositStatus } from "./poaBridgeHttpClient"
 
 type TokenBalances = Record<BaseTokenInfo["defuseAssetId"], bigint>
 
@@ -26,26 +24,4 @@ export async function getDepositedBalances(
   }
 
   return result
-}
-
-export async function getPendingDeposits(
-  accountId: IntentsUserId,
-  tokenIds: BaseTokenInfo["defuseAssetId"][]
-): Promise<TokenBalances> {
-  const depositStatus = await getDepositStatus({
-    account_id: accountId,
-  })
-
-  const pendingDeposits: TokenBalances = {}
-
-  for (const deposit of depositStatus.deposits) {
-    // POA bridge returns token IDs without the 'nep141:' prefix (e.g. 'base.omft.near')
-    const defuseAssetId = tokenAccountIdToDefuseAssetId(deposit.near_token_id)
-
-    if (tokenIds.includes(defuseAssetId) && deposit.status === "PENDING") {
-      pendingDeposits[defuseAssetId] = BigInt(deposit.amount)
-    }
-  }
-
-  return pendingDeposits
 }
