+++ src/services/poaBridgeService.ts
@@ -1,7 +1,14 @@
 import { RpcRequestError } from "../errors/request"
+import type { BaseTokenInfo } from "../types/base"
+import type { IntentsUserId } from "../types/intentsUserId"
 import { assert, type AssertErrorType } from "../utils/assert"
+import { tokenAccountIdToDefuseAssetId } from "../utils/tokenUtils"
 import { wait } from "../utils/wait"
-import { getWithdrawalStatus, type types } from "./poaBridgeHttpClient"
+import {
+  getDepositStatus,
+  getWithdrawalStatus,
+  type types,
+} from "./poaBridgeHttpClient"
 
 export type WaitForWithdrawalCompletionOkType = {
   destinationTxHash: string
@@ -62,4 +69,33 @@ function isWithdrawalNotFound(err: unknown) {
     err instanceof RpcRequestError &&
     err.details === RPC_ERR_MSG_WITHDRAWALS_NOT_FOUND
   )
+}
+
+export type GetPendingDepositsOkType = Record<
+  BaseTokenInfo["defuseAssetId"],
+  bigint
+>
+
+export type GetPendingDepositsErrorType = types.JSONRPCErrorType
+
+export async function getPendingDeposits(
+  accountId: IntentsUserId,
+  tokenIds: BaseTokenInfo["defuseAssetId"][]
+): Promise<GetPendingDepositsOkType> {
+  const depositStatus = await getDepositStatus({
+    account_id: accountId,
+  })
+
+  const pendingDeposits: GetPendingDepositsOkType = {}
+
+  for (const deposit of depositStatus.deposits) {
+    // POA bridge returns token IDs without the 'nep141:' prefix (e.g. 'base.omft.near')
+    const defuseAssetId = tokenAccountIdToDefuseAssetId(deposit.near_token_id)
+
+    if (tokenIds.includes(defuseAssetId) && deposit.status === "PENDING") {
+      pendingDeposits[defuseAssetId] = BigInt(deposit.amount)
+    }
+  }
+
+  return pendingDeposits
 }
